Added admin Functionality

+++ lib/actions.ts
@@ -1,12 +1,14 @@
 "use server";
 
-import {blogschema, commentschema, deletebtn, publisherschema, saveblogschema} from "@/lib/zodschema";
+import {blogschema, commentschema, deletebtn, deletecategory, publisherschema, saveblogschema} from "@/lib/zodschema";
 import {actionClient} from "@/lib/safe-action";
 import {revalidatePath} from "next/cache";
 import {currentUser} from "@clerk/nextjs/server";
 import {eq} from "drizzle-orm";
 import {db} from "@/app/db/db";
-import {comments, posts, users} from "@/app/db/schema";
+import {categories, comments, posts, users} from "@/app/db/schema";
+import adminControl, { Allowedtodelete } from "./AdminControl";
+import { redirect } from "next/navigation";
 
 
 export async function getUserIdAction() {
@@ -36,25 +38,28 @@ export const createcomment = actionClient.schema(commentschema).action(async ({ 
         return {failure: "Incorrect credentials"};
     }
 });
-export const createBlogAction = actionClient
-    .schema(blogschema)
-    .action(async ({ parsedInput: { title, slug, content, headerimageurl, description,categoryid } }) => {
+export const createBlogAction = actionClient.schema(blogschema).action(async ({ parsedInput: { title, slug, content, headerimageurl, description,categoryid } }) => {
         try {
             const userid = await getUserIdAction()
-            await db.insert(posts).values({
-                title: title,
-                slug: slug,
-                content: content,
-                headerimageurl: headerimageurl,
-                authorId: userid,
-                categoryId: categoryid,
-                isfeatured: false,
-                ispublished: true,
-                description: description
-            })
-            return {
-                success: 'Added Successfully',
-            };
+            if(userid){
+                await db.insert(posts).values({
+                    title: title,
+                    slug: slug,
+                    content: content,
+                    headerimageurl: headerimageurl,
+                    authorId: userid,
+                    categoryId: categoryid,
+                    isfeatured: false,
+                    ispublished: true,
+                    description: description
+                })
+                redirect(`/dashboard`)
+                return {
+                    success: 'Added Successfully',
+                };
+            }else{
+                return { failure: "Incorrect Data" };
+            }
         }catch (error: any) {
             return { failure: "Incorrect credentials" };
         }
@@ -116,14 +121,37 @@ export const publishBlogAction = actionClient.schema(publisherschema).action(asy
 export const DeleteBlogAction = actionClient.schema(deletebtn).action(async ({ parsedInput: { postId } }) => {
     try {
         const userid = await getUserIdAction()
-        await db.delete(comments).where(eq(comments.postId, postId))
-        await db.delete(posts).where(eq(posts.id, postId))
-        revalidatePath(`/dashboard`)
-        return {
-            success: 'Deleted Successfully',
-        };
+        const isallowed = await Allowedtodelete(userid)
+        if(isallowed){
+            await db.delete(comments).where(eq(comments.postId, postId))
+            await db.delete(posts).where(eq(posts.id, postId))
+            revalidatePath(`/dashboard`)
+            return {
+                success: 'Deleted Successfully',
+            };
+        }else {
+            return { failure: "Incorrect credentials" };
+        }
     }catch (error: any) {
         return { failure: "Incorrect credentials" };
     }
 
 });
+
+export const deletecategoryAction = actionClient.schema(deletecategory).action(async({parsedInput: {categoryId}}) => {
+    try{
+        const isadmin = await adminControl()
+        if(isadmin){
+            await db.delete(posts).where(eq(posts.categoryId, categoryId))
+            await db.delete(categories).where(eq(categories.id, categoryId))
+            revalidatePath(`/admin`)
+            return {
+                success: 'Deleted Successfully',
+            };
+        }else {
+            return { failure: "Incorrect credentials" };
+        }
+    }catch (error: any) {
+        return { failure: "Incorrect credentials" };
+    }
+})
